refactor: convert admin pages to RSCs
