+++ apps/web/src/app/admin/login/page.tsx
@@ -1,135 +1,14 @@
-"use client";
+import { LoginForm } from "@web/app/admin/components/login-form";
 
-import { Button } from "@sgcarstrends/ui/components/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@sgcarstrends/ui/components/card";
-import { Input } from "@sgcarstrends/ui/components/input";
-import { Label } from "@sgcarstrends/ui/components/label";
-import { authClient } from "@web/app/admin/lib/auth-client";
-import { useRouter } from "next/navigation";
-import { type FormEvent, useState } from "react";
-import { toast } from "sonner";
-
-const LoginPage = () => {
-  const router = useRouter();
-  const [isLoading, setIsLoading] = useState(false);
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-
-    await authClient.signIn.email(
-      {
-        email,
-        password,
-      },
-      {
-        onSuccess: () => {
-          toast.success("Signed in successfully");
-          router.push("/admin");
-          router.refresh();
-        },
-        onError: (ctx) => {
-          toast.error(ctx.error.message || "Failed to sign in");
-          setIsLoading(false);
-        },
-      },
-    );
-  };
-
-  const handleGoogleSignIn = async () => {
-    setIsLoading(true);
-    await authClient.signIn.social({
-      provider: "google",
-      callbackURL: "/admin",
-    });
-  };
-
+export default function LoginPage() {
   return (
     <div className="flex min-h-svh flex-col items-center justify-center gap-6 bg-muted p-6 md:p-10">
       <div className="flex w-full max-w-sm flex-col gap-6">
-        <Card>
-          <CardHeader className="text-center">
-            <CardTitle className="text-xl">Welcome back</CardTitle>
-            <CardDescription>
-              Login to access the admin dashboard
-            </CardDescription>
-          </CardHeader>
-          <CardContent>
-            <form onSubmit={handleSubmit} className="flex flex-col gap-6">
-              <div className="flex flex-col gap-4">
-                <Button
-                  variant="outline"
-                  type="button"
-                  onClick={handleGoogleSignIn}
-                  disabled={isLoading}
-                >
-                  <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    viewBox="0 0 24 24"
-                    className="size-4"
-                    aria-label="Google logo"
-                  >
-                    <title>Google</title>
-                    <path
-                      d="M12.48 10.92v3.28h7.84c-.24 1.84-.853 3.187-1.787 4.133-1.147 1.147-2.933 2.4-6.053 2.4-4.827 0-8.6-3.893-8.6-8.72s3.773-8.72 8.6-8.72c2.6 0 4.507 1.027 5.907 2.347l2.307-2.307C18.747 1.44 16.133 0 12.48 0 5.867 0 .307 5.387.307 12s5.56 12 12.173 12c3.573 0 6.267-1.173 8.373-3.36 2.16-2.16 2.84-5.213 2.84-7.667 0-.76-.053-1.467-.173-2.053H12.48z"
-                      fill="currentColor"
-                    />
-                  </svg>
-                  Login with Google
-                </Button>
-              </div>
-
-              <div className="relative text-center text-sm after:absolute after:inset-0 after:top-1/2 after:z-0 after:flex after:items-center after:border-border after:border-t">
-                <span className="relative z-10 bg-card px-2 text-muted-foreground">
-                  Or continue with
-                </span>
-              </div>
-
-              <div className="flex flex-col gap-4">
-                <div className="flex flex-col gap-2">
-                  <Label htmlFor="email">Email</Label>
-                  <Input
-                    id="email"
-                    type="email"
-                    placeholder="m@example.com"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    required
-                    disabled={isLoading}
-                  />
-                </div>
-                <div className="flex flex-col gap-2">
-                  <Label htmlFor="password">Password</Label>
-                  <Input
-                    id="password"
-                    type="password"
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                    required
-                    disabled={isLoading}
-                  />
-                </div>
-                <Button type="submit" disabled={isLoading}>
-                  {isLoading ? "Signing in..." : "Login"}
-                </Button>
-              </div>
-            </form>
-          </CardContent>
-        </Card>
+        <LoginForm />
         <div className="text-balance px-6 text-center text-muted-foreground text-xs">
           Protected admin area - Authorised users only
         </div>
       </div>
     </div>
   );
-};
-
-export default LoginPage;
+}
