+++ apps/web/src/proxy.ts
@@ -1,7 +1,7 @@
 import crypto from "node:crypto";
 import { db, sessions } from "@sgcarstrends/database";
 import { redis } from "@sgcarstrends/utils";
-import { auth } from "@web/app/admin/_lib/auth";
+import { auth } from "@web/app/admin/lib/auth";
 import { and, eq, gt } from "drizzle-orm";
 import { headers } from "next/headers";
 import { type NextRequest, NextResponse } from "next/server";
